+++ app/controllers/AvatarController.php
@@ -1,0 +1,11 @@
+<?php
+
+use Movim\Controller\Base;
+
+class AvatarController extends Base
+{
+    public function load()
+    {
+        $this->raw = true;
+    }
+}