+++ app/controllers/ColorsController.php
@@ -1,0 +1,10 @@
+<?php
+use Movim\Controller\Base;
+
+class ColorsController extends Base
+{
+    public function load()
+    {
+        $this->raw = true;
+    }
+}